# [ready] perf: create buffer only when needed