@tinyb0t 2023-08-26
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/lazy.py                           4               7
------------------------------------------------------------
lines added in the tinygrad folder: -3
```