@geohot 2023-08-31
So this is actually a good change. But it should be a 2 line diff!

Always think what's the minimal change to achieve the outcome. There's two unrelated formatting changes in here.