@roelofvandijk 2023-08-31
Minimized diff. There are other (minor) changes that can make this faster, but this is the big one.